macosx: explicitly show the protocols the objects are implementing, use a modern API method in the wizard and fixed some minor warnings

+++ modules/gui/macosx/playlistinfo.h
@@ -126,8 +126,8 @@
     NSMutableArray *o_children;
 }
 
-- (int)numberOfChildren;
-- (VLCInfoTreeItem *)childAtIndex:(int)i_index;
+- (NSInteger)numberOfChildren;
+- (VLCInfoTreeItem *)childAtIndex:(NSUInteger)i_index;
 - (NSString *)name;
 - (NSString *)value;
 - (void)refresh;
